--- src/test/java/ladder/domain/HeightTest.java
@@ -1,16 +1,0 @@
-package ladder.domain;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-
-class HeightTest {
-
-    @DisplayName("입력 값이 2보다 작으면 예외 발생")
-    @Test
-    void valueOfInvalidRange() {
-        assertThatIllegalArgumentException()
-                .isThrownBy(() -> Height.valueOf(1));
-    }
-}
