+++ src/test/java/ladder/domain/player/PlayerTest.java
@@ -1,5 +1,7 @@
-package ladder.domain;
+package ladder.domain.player;
 
+import ladder.domain.player.Name;
+import ladder.domain.player.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
